Added Commandhandler Changed id from Long to int

+++ src/main/java/to/mattias/controllers/ProjectController.java
@@ -2,6 +2,7 @@ package to.mattias.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import to.mattias.beans.Cmd;
 import to.mattias.entities.Project;
 import to.mattias.services.ProjectService;
 
@@ -28,7 +29,7 @@ public class ProjectController {
     }
 
     @RequestMapping(value = "/{projectId}", method = RequestMethod.GET)
-    public Project findById(@PathVariable Long projectId) {
+    public Project findById(@PathVariable int projectId) {
         return service.findById(projectId);
     }
 
@@ -39,7 +40,7 @@ public class ProjectController {
 
     @DeleteMapping
     @RequestMapping("/{projectId}")
-    public void delete(@PathVariable Long projectId) {
+    public void delete(@PathVariable int projectId) {
         service.delete(projectId);
     }
 }
