+++ src/main/java/to/mattias/repositories/NoteObjRepository.java
@@ -6,5 +6,5 @@ import to.mattias.entities.NoteObj;
 /**
  * <h1>Created by Mattias on 2017-01-23.</h1>
  */
-public interface NoteObjRepository extends JpaRepository<NoteObj, Long> {
+public interface NoteObjRepository extends JpaRepository<NoteObj, Integer> {
 }
