+++ src/main/java/to/mattias/repositories/SprintRepository.java
@@ -6,6 +6,6 @@ import to.mattias.entities.Sprint;
 /**
  * <h1>Created by Mattias on 2017-01-23.</h1>
  */
-public interface SprintRepository extends JpaRepository<Sprint, Long> {
+public interface SprintRepository extends JpaRepository<Sprint, Integer> {
 
 }
