+++ src/main/java/to/mattias/repositories/TaskRepository.java
@@ -6,5 +6,5 @@ import to.mattias.entities.Task;
 /**
  * <h1>Created by Mattias on 2017-01-23.</h1>
  */
-public interface TaskRepository extends JpaRepository<Task, Long> {
+public interface TaskRepository extends JpaRepository<Task, Integer> {
 }
